# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FacturaloYa",
        "description": "Sistema de facturacion en linea. Con carrito de compras",
        "contact": {
            "email": "neisserrey@upeu.edu.pe"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://facturacionmvc.test/api/",
            "description": "Documentacion de la Facturacion en Linea"
        }
    ],
    "paths": {
        "/categories": {
            "get": {
                "tags": [
                    "categories"
                ],
                "summary": "Mostrar categorias",
                "operationId": "App\\Http\\Controllers\\Category\\CategoryController::index",
                "responses": {
                    "401": {
                        "description": "Inserta tu token pues hermano!"
                    },
                    "200": {
                        "description": "Lista de categorias. Correcto"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "categories"
                ],
                "summary": "Agregar una nueva categoria.",
                "operationId": "store",
                "requestBody": {
                    "description": "Categoria a ser creada",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Category"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Inserta tu token pues hermano!"
                    },
                    "201": {
                        "description": "Creado"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/categories/{categoryId}": {
            "get": {
                "tags": [
                    "categories"
                ],
                "summary": "Buscar categoria por ID",
                "description": "Retorna una sola categoria",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "categoryId",
                        "in": "path",
                        "description": "ID de la categoria a retornar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Inserta tu token pues hermano!"
                    },
                    "404": {
                        "description": "Categoria no existe"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "categories"
                ],
                "summary": "Elimina una categoria",
                "description": "",
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "categoryId",
                        "in": "path",
                        "description": "Id de la categoria a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Categoria no encontrada"
                    },
                    "401": {
                        "description": "Inserta tu token pues hermano!"
                    },
                    "403": {
                        "description": "Tiene productos relacionados"
                    },
                    "200": {
                        "description": "Categoria eliminada correctamente"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/category/{categoryId}": {
            "put": {
                "tags": [
                    "category"
                ],
                "summary": "Actualizar una categoria existente",
                "description": "",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "categoryId",
                        "in": "path",
                        "description": "ID de la categoria a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Categoria a ser actualizada",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Category"
                            }
                        }
                    }
                },
                "responses": {
                    "404": {
                        "description": "Categoria no encontrada"
                    },
                    "401": {
                        "description": "Inserta tu token pues hermano!"
                    },
                    "200": {
                        "description": "Categoria actualizada correctamente"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Category": {
                "properties": {
                    "name": {
                        "description": "Nombre de la categoria",
                        "type": "string"
                    },
                    "description": {
                        "description": "Descripcion de la categoria",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Authorization token obtained from logging in.",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}